+++ src/api-wrappers/CGWindow.swift
@@ -4,18 +4,11 @@ typealias CGWindow = [CFString: Any]
 
 extension CGWindow {
     static let normalLevel = CGWindowLevelForKey(.normalWindow)
+    static let baseLevel = CGWindowLevelForKey(.baseWindow)
+    static let floatingWindow = CGWindowLevelForKey(.floatingWindow)
 
     static func windows(_ option: CGWindowListOption) -> [CGWindow] {
         return CGWindowListCopyWindowInfo([.excludeDesktopElements, option], kCGNullWindowID) as! [CGWindow]
-    }
-
-    static func isMissionControlActive() -> Bool {
-        // when Mission Control is active, the Dock process spawns some windows. We observe this side-effect and infer
-        for window in windows(.optionOnScreenOnly) {
-            guard window.ownerName() == "Dock", window.title() == nil else { continue }
-            return true
-        }
-        return false
     }
 
     // workaround: filtering this criteria seems to remove non-windows UI elements