+++ src/logic/Windows.swift
@@ -10,18 +10,39 @@ class Windows {
     /// reordered list based on preferences, keeping the original index
     static func reorderList() {
         list.sort {
-            if let bool = sortByBooleanAttribute($0.isWindowlessApp, $1.isWindowlessApp) {
-                return bool
+            // separate buckets for these types of windows
+            if $0.isWindowlessApp != $1.isWindowlessApp {
+                return $1.isWindowlessApp
             }
-            if Preferences.showHiddenWindows[App.app.shortcutIndex] == .showAtTheEnd,
-               let bool = sortByBooleanAttribute($0.isHidden, $1.isHidden) {
-                return bool
+            if Preferences.showHiddenWindows[App.app.shortcutIndex] == .showAtTheEnd && $0.isHidden != $1.isHidden {
+               return $1.isHidden
             }
-            if Preferences.showMinimizedWindows[App.app.shortcutIndex] == .showAtTheEnd,
-               let bool = sortByBooleanAttribute($0.isMinimized, $1.isMinimized) {
-                return bool
+            if Preferences.showMinimizedWindows[App.app.shortcutIndex] == .showAtTheEnd && $0.isMinimized != $1.isMinimized {
+                return $1.isMinimized
             }
-            return $0.lastFocusOrder < $1.lastFocusOrder
+
+            // sort within each buckets
+            let sortType = Preferences.windowOrder[App.app.shortcutIndex]
+            if sortType == .recentlyFocused {
+                return $0.lastFocusOrder < $1.lastFocusOrder
+            }
+            if sortType == .recentlyCreated {
+                return $1.creationOrder < $0.creationOrder
+            }
+            var order = ComparisonResult.orderedSame
+            if sortType == .alphabetical {
+                order = sortByAppNameThenWindowTitle($0, $1)
+            }
+            if sortType == .space {
+                order = $0.spaceIndex.compare($1.spaceIndex)
+                if order == .orderedSame {
+                    order = sortByAppNameThenWindowTitle($0, $1)
+                }
+            }
+            if order == .orderedSame {
+                order = $0.lastFocusOrder.compare($1.lastFocusOrder)
+            }
+            return order == .orderedAscending
         }
         if (DockAltTabDockPos == "right") {list = list.reversed()}
     }
@@ -35,7 +56,7 @@ class Windows {
             ThumbnailsView.highlight(oldIndex)
         }
         if let app = Applications.find(NSWorkspace.shared.frontmostApplication?.processIdentifier),
-           app.focusedWindow == nil,
+           (app.focusedWindow == nil || Preferences.windowOrder[App.app.shortcutIndex] != .recentlyFocused),
            let lastFocusedWindowIndex = getLastFocusedWindowIndex() {
             updateFocusedAndHoveredWindowIndex(lastFocusedWindowIndex)
         } else {
@@ -128,7 +149,7 @@ class Windows {
         App.app.thumbnailsPanel.thumbnailsView.scrollView.contentView.scrollToVisible(focusedView.frame)
         voiceOverWindow(index)
     }
-    
+
     static func previewFocusedWindowIfNeeded() {
         guard
             Preferences.previewFocusedWindow,
@@ -195,6 +216,8 @@ class Windows {
         if let focusedWindow = focusedWindow() {
             if !focusedWindow.shouldShowTheUser {
                 cycleFocusedWindowIndex(windowIndexAfterCycling(1) > focusedWindowIndex ? 1 : -1)
+            } else {
+                previewFocusedWindowIfNeeded()
             }
         } else {
             cycleFocusedWindowIndex(-1)
@@ -211,10 +234,21 @@ class Windows {
     /// tabs detection is a flaky work-around the lack of public API to observe OS tabs
     /// see: https://github.com/lwouis/alt-tab-macos/issues/1540
     static func detectTabbedWindows() {
-        let cgsWindowIds = Spaces.windowsInSpaces(Spaces.idsAndIndexes.map { $0.0 })
+        lazy var cgsWindowIds = Spaces.windowsInSpaces(Spaces.idsAndIndexes.map { $0.0 })
+        lazy var visibleCgsWindowIds = Spaces.windowsInSpaces(Spaces.idsAndIndexes.map { $0.0 }, false)
         list.forEach {
             if let cgWindowId = $0.cgWindowId {
-                $0.isTabbed = !cgsWindowIds.contains(cgWindowId)
+                if $0.isMinimized || $0.isHidden {
+                    if #available(macOS 13, *) {
+                        // not exact after window merging
+                        $0.isTabbed = !cgsWindowIds.contains(cgWindowId)
+                    } else {
+                        // not known
+                        $0.isTabbed = false
+                    }
+                } else {
+                    $0.isTabbed = !visibleCgsWindowIds.contains(cgWindowId)
+                }
             }
         }
     }
@@ -289,13 +323,10 @@ class Windows {
     }
 }
 
-func sortByBooleanAttribute(_ b1: Bool, _ b2: Bool) -> Bool? {
-    if b1 && !b2 {
-        return false
+func sortByAppNameThenWindowTitle(_ w1: Window, _ w2: Window) -> ComparisonResult {
+    var order = w1.application.runningApplication.localizedName.localizedStandardCompare(w2.application.runningApplication.localizedName)
+    if order == .orderedSame {
+        return w1.title.localizedStandardCompare(w2.title)
     }
-    if !b1 && b2 {
-        return true
-    }
-    return nil
+    return order
 }
-