+++ src/ui/App.swift
@@ -16,6 +16,7 @@ class App: AppCenterApplication, NSApplicationDelegate {
     static let licence = Bundle.main.object(forInfoDictionaryKey: "NSHumanReadableCopyright") as! String
     static let repository = "https://github.com/steventheworker/alt-tab-macos"
     static var app: App!
+    var menubar: Menubar!
     var thumbnailsPanel: ThumbnailsPanel!
     var previewPanel: PreviewPanel!
     var preferencesWindow: PreferencesWindow!
@@ -53,7 +54,7 @@ class App: AppCenterApplication, NSApplicationDelegate {
             guard let self = self else { return }
             BackgroundWork.start()
             Preferences.initialize()
-            Menubar.initialize()
+            self.menubar = Menubar()
             self.loadMainMenuXib()
             self.thumbnailsPanel = ThumbnailsPanel()
             self.previewPanel = PreviewPanel()
@@ -109,13 +110,15 @@ class App: AppCenterApplication, NSApplicationDelegate {
         App.shared.terminate(self)
     }
 
-    func hideUi() {
+    func hideUi(_ keepPreview: Bool = false) {
         debugPrint("hideUi")
         appIsBeingUsed = false
         isFirstSummon = true
         MouseEvents.toggle(false)
         hideThumbnailPanelWithoutChangingKeyWindow()
-        previewPanel.orderOut(nil)
+        if !keepPreview {
+            previewPanel.orderOut(nil)
+        }
         hideAllTooltips()
     }
 
@@ -142,6 +145,10 @@ class App: AppCenterApplication, NSApplicationDelegate {
 
     func minDeminSelectedWindow() {
         Windows.focusedWindow()?.minDemin()
+    }
+
+    func toggleFullscreenSelectedWindow() {
+        Windows.focusedWindow()?.toggleFullscreen()
     }
 
     func quitSelectedApp() {
@@ -201,11 +208,14 @@ class App: AppCenterApplication, NSApplicationDelegate {
     }
 
     func focusSelectedWindow(_ selectedWindow: Window?) {
-        hideUi()
-        guard let window = selectedWindow, !CGWindow.isMissionControlActive() else { return }
-        window.focus()
-        if Preferences.cursorFollowFocusEnabled {
-            moveCursorToSelectedWindow(window)
+        hideUi(true)
+        if let window = selectedWindow, !MissionControl.isActive() {
+            window.focus()
+            if Preferences.cursorFollowFocusEnabled {
+                moveCursorToSelectedWindow(window)
+            }
+        } else {
+            previewPanel.orderOut(nil)
         }
     }
 
@@ -259,7 +269,7 @@ class App: AppCenterApplication, NSApplicationDelegate {
         if isFirstSummon || shortcutIndex != self.shortcutIndex {
             debugPrint("showUiOrCycleSelection: isFirstSummon")
             isFirstSummon = false
-            if Windows.list.count == 0 || CGWindow.isMissionControlActive() { hideUi(); return }
+            if Windows.list.count == 0 || MissionControl.isActive() { hideUi(); return }
             // TODO: can the CGS call inside detectTabbedWindows introduce latency when WindowServer is busy?
             Windows.detectTabbedWindows()
             // TODO: find a way to update space info when spaces are changed, instead of on every trigger